no concerns dir, rename table test trait
